feat: implement CRUD for FoodEntries

+++ FoodTracker.Api/Contracts/PatchFoodItemRequest.cs
@@ -5,8 +5,8 @@ public sealed class PatchFoodItemRequest
     public string? Name { get; set; }
     public string? Category { get; set; }
     public string? Description { get; set; }
-    public decimal? CaloriesPer100g { get; set; }
-    public decimal? ProteinsPer100g { get; set; }
-    public decimal? FatsPer100g { get; set; }
-    public decimal? CarbsPer100g { get; set; }
+    public double? CaloriesPer100g { get; set; }
+    public double? ProteinsPer100g { get; set; }
+    public double? FatsPer100g { get; set; }
+    public double? CarbsPer100g { get; set; }
 }
