+++ FoodTracker.Api/Contracts/UpdateFoodItemRequest.cs
@@ -4,8 +4,8 @@ public sealed class UpdateFoodItemRequest
 {
     public string Name { get; init; } = string.Empty;
     public string? Description { get; init; }
-    public decimal CaloriesPer100g { get; init; }
-    public decimal ProteinsPer100g { get; init; }
-    public decimal FatsPer100g { get; init; }
-    public decimal CarbsPer100g { get; init; }
+    public double CaloriesPer100g { get; init; }
+    public double ProteinsPer100g { get; init; }
+    public double FatsPer100g { get; init; }
+    public double CarbsPer100g { get; init; }
 }
