+++ FoodTracker.Application/Features/Nutrition/PatchFoodItemCommand.cs
@@ -10,8 +10,8 @@ public sealed class PatchFoodItemCommand : IRequest<Result<FoodItemDto>>
     public Guid FoodItemId { get; init; }
     public string? Name { get; init; }
     public string? Description { get; init; }
-    public decimal? CaloriesPer100g { get; init; }
-    public decimal? ProteinsPer100g { get; init; }
-    public decimal? FatsPer100g { get; init; }
-    public decimal? CarbsPer100g { get; init; }
+    public double? CaloriesPer100g { get; init; }
+    public double? ProteinsPer100g { get; init; }
+    public double? FatsPer100g { get; init; }
+    public double? CarbsPer100g { get; init; }
 }
